requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
django-environ==0.4.5
Flask==1.1.1
googlemaps==3.1.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
nltk==3.4.5
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pytest==5.2.2
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wikipedia==1.4.0
zipp==0.6.0
